lang/c: snappy has no pkg-config entry
Google explicitly is not supporting that:

    https://github.com/google/snappy/pull/86

Which leaves us users in a somewhat precarious state; if users
of snappy, like avro, try to include it as a requirement in
their pkg-config, they need to start special-casing for each
possible packager.

Currently `libsnappy` ends up in the generated `Requires` section,
which causes any use of `pkg-config --exists avro-c` to return
a non-zero status on a huge chunk of platforms.

This commit removes `libsnappy` from the Requires.